container/types: reimplement Packer with malloc (#17702)
replace mpool allocator with malloc allocator in Packer

Approved by: @badboynt1, @ouyuanning, @XuPeng-SH, @triump2020, @m-schen, @aunjgr
