+++ pkg/container/types/tuple.go
@@ -31,12 +31,9 @@ import (
 	"bytes"
 	"encoding/binary"
 	"fmt"
-	"math"
 	"strconv"
 	"strings"
 	"unsafe"
-
-	"github.com/matrixorigin/matrixone/pkg/common/mpool"
 
 	"github.com/matrixorigin/matrixone/pkg/common/moerr"
 )
@@ -240,279 +237,6 @@ func adjustFloatBytes(b []byte, encode bool) {
 		// Positive number: flip just the sign bit.
 		b[0] = b[0] ^ 0x80
 	}
-}
-
-const PackerMemUnit = 64
-
-type Packer struct {
-	buf      []byte
-	size     int
-	capacity int
-	mp       *mpool.MPool
-}
-
-func NewPacker(mp *mpool.MPool) *Packer {
-	bytes, err := mp.Alloc(PackerMemUnit)
-	if err != nil {
-		panic(err)
-	}
-	return &Packer{
-		buf:      bytes,
-		size:     0,
-		capacity: PackerMemUnit,
-		mp:       mp,
-	}
-}
-
-func NewPackerArray(length int, mp *mpool.MPool) []*Packer {
-	packerArr := make([]*Packer, length)
-	for num := range packerArr {
-		bytes, err := mp.Alloc(PackerMemUnit)
-		if err != nil {
-			panic(err)
-		}
-		packerArr[num] = &Packer{
-			buf:      bytes,
-			size:     0,
-			capacity: PackerMemUnit,
-			mp:       mp,
-		}
-	}
-	return packerArr
-}
-
-func (p *Packer) FreeMem() {
-	if p.buf != nil {
-		p.mp.Free(p.buf)
-		p.size = 0
-		p.capacity = 0
-		p.buf = nil
-	}
-}
-
-func (p *Packer) Reset() {
-	p.size = 0
-}
-
-func (p *Packer) putByte(b byte) {
-	if p.size < p.capacity {
-		p.buf[p.size] = b
-		p.size++
-	} else {
-		p.buf, _ = p.mp.Grow(p.buf, p.capacity+PackerMemUnit)
-		p.capacity += PackerMemUnit
-		p.buf[p.size] = b
-		p.size++
-	}
-}
-
-func (p *Packer) putBytes(bs []byte) {
-	if p.size+len(bs) < p.capacity {
-		for _, b := range bs {
-			p.buf[p.size] = b
-			p.size++
-		}
-	} else {
-		incrementSize := ((len(bs) / PackerMemUnit) + 1) * PackerMemUnit
-		p.buf, _ = p.mp.Grow(p.buf, p.capacity+incrementSize)
-		p.capacity += incrementSize
-		for _, b := range bs {
-			p.buf[p.size] = b
-			p.size++
-		}
-	}
-}
-
-func (p *Packer) putBytesNil(b []byte, i int) {
-	for i >= 0 {
-		p.putBytes(b[:i+1])
-		p.putByte(0xFF)
-		b = b[i+1:]
-		i = bytes.IndexByte(b, 0x00)
-	}
-	p.putBytes(b)
-}
-
-func (p *Packer) encodeBytes(code byte, b []byte) {
-	p.putByte(code)
-	if i := bytes.IndexByte(b, 0x00); i >= 0 {
-		p.putBytesNil(b, i)
-	} else {
-		p.putBytes(b)
-	}
-	p.putByte(0x00)
-}
-
-func (p *Packer) encodeUint(i uint64) {
-	if i == 0 {
-		p.putByte(intZeroCode)
-		return
-	}
-
-	n := bisectLeft(i)
-	var scratch [8]byte
-
-	p.putByte(byte(intZeroCode + n))
-	binary.BigEndian.PutUint64(scratch[:], i)
-
-	p.putBytes(scratch[8-n:])
-}
-
-func (p *Packer) encodeInt(i int64) {
-	if i >= 0 {
-		p.encodeUint(uint64(i))
-		return
-	}
-
-	n := bisectLeft(uint64(-i))
-	var scratch [8]byte
-
-	p.putByte(byte(intZeroCode - n))
-	offsetEncoded := int64(sizeLimits[n]) + i
-	binary.BigEndian.PutUint64(scratch[:], uint64(offsetEncoded))
-
-	p.putBytes(scratch[8-n:])
-}
-
-func (p *Packer) encodeFloat32(f float32) {
-	var scratch [4]byte
-	binary.BigEndian.PutUint32(scratch[:], math.Float32bits(f))
-	adjustFloatBytes(scratch[:], true)
-
-	p.putByte(float32Code)
-	p.putBytes(scratch[:])
-}
-
-func (p *Packer) encodeFloat64(d float64) {
-	var scratch [8]byte
-	binary.BigEndian.PutUint64(scratch[:], math.Float64bits(d))
-	adjustFloatBytes(scratch[:], true)
-
-	p.putByte(float64Code)
-	p.putBytes(scratch[:])
-}
-
-func (p *Packer) EncodeInt8(e int8) {
-	p.putByte(int8Code)
-	p.encodeInt(int64(e))
-}
-
-func (p *Packer) EncodeInt16(e int16) {
-	p.putByte(int16Code)
-	p.encodeInt(int64(e))
-}
-
-func (p *Packer) EncodeInt32(e int32) {
-	p.putByte(int32Code)
-	p.encodeInt(int64(e))
-}
-
-func (p *Packer) EncodeInt64(e int64) {
-	p.putByte(int64Code)
-	p.encodeInt(e)
-}
-
-func (p *Packer) EncodeUint8(e uint8) {
-	p.putByte(uint8Code)
-	p.encodeUint(uint64(e))
-}
-
-func (p *Packer) EncodeUint16(e uint16) {
-	p.putByte(uint16Code)
-	p.encodeUint(uint64(e))
-}
-
-func (p *Packer) EncodeUint32(e uint32) {
-	p.putByte(uint32Code)
-	p.encodeUint(uint64(e))
-}
-
-func (p *Packer) EncodeUint64(e uint64) {
-	p.putByte(uint64Code)
-	p.encodeUint(e)
-}
-
-func (p *Packer) EncodeFloat32(e float32) {
-	p.encodeFloat32(e)
-}
-
-func (p *Packer) EncodeFloat64(e float64) {
-	p.encodeFloat64(e)
-}
-
-func (p *Packer) EncodeNull() {
-	p.putByte(nilCode)
-}
-
-func (p *Packer) EncodeBool(e bool) {
-	if e {
-		p.putByte(trueCode)
-	} else {
-		p.putByte(falseCode)
-	}
-}
-
-func (p *Packer) EncodeDate(e Date) {
-	p.putByte(dateCode)
-	p.encodeInt(int64(e))
-}
-
-func (p *Packer) EncodeTime(e Time) {
-	p.putByte(timeCode)
-	p.encodeInt(int64(e))
-}
-
-func (p *Packer) EncodeDatetime(e Datetime) {
-	p.putByte(datetimeCode)
-	p.encodeInt(int64(e))
-}
-
-func (p *Packer) EncodeTimestamp(e Timestamp) {
-	p.putByte(timestampCode)
-	p.encodeInt(int64(e))
-}
-
-func (p *Packer) EncodeEnum(e Enum) {
-	p.putByte(enumCode)
-	p.EncodeUint16(uint16(e))
-}
-
-func (p *Packer) EncodeDecimal64(e Decimal64) {
-	p.putByte(decimal64Code)
-	b := *(*[8]byte)(unsafe.Pointer(&e))
-	b[7] ^= 0x80
-	for i := 7; i >= 0; i-- {
-		p.putByte(b[i])
-	}
-}
-
-func (p *Packer) EncodeDecimal128(e Decimal128) {
-	p.putByte(decimal128Code)
-	b := *(*[16]byte)(unsafe.Pointer(&e))
-	b[15] ^= 0x80
-	for i := 15; i >= 0; i-- {
-		p.putByte(b[i])
-	}
-}
-
-func (p *Packer) EncodeStringType(e []byte) {
-	p.putByte(stringTypeCode)
-	p.encodeBytes(bytesCode, e)
-}
-
-func (p *Packer) EncodeBit(e uint64) {
-	p.putByte(bitCode)
-	p.encodeUint(e)
-}
-
-func (p *Packer) GetBuf() []byte {
-	return p.buf[:p.size]
-}
-
-func (p *Packer) Bytes() []byte {
-	ret := make([]byte, p.size)
-	copy(ret, p.buf[:p.size])
-	return ret
 }
 
 func findTerminator(b []byte) int {
