+++ pkg/sql/colexec/lockop/lock_op_no_txn.go
@@ -59,8 +59,8 @@ func LockTableWithUniqueID(
 		return err
 	}
 
-	parker := types.NewPacker(proc.Mp())
-	defer parker.FreeMem()
+	parker := types.NewPacker()
+	defer parker.Close()
 
 	opts := DefaultLockOptions(parker).
 		WithLockMode(mode).
