+++ pkg/vm/engine/disttae/txn_table_test.go
@@ -29,20 +29,18 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func newTxnTableForTest(
-	mp *mpool.MPool,
-) *txnTable {
+func newTxnTableForTest() *txnTable {
 	engine := &Engine{
 		packerPool: fileservice.NewPool(
 			128,
 			func() *types.Packer {
-				return types.NewPacker(mp)
+				return types.NewPacker()
 			},
 			func(packer *types.Packer) {
 				packer.Reset()
 			},
 			func(packer *types.Packer) {
-				packer.FreeMem()
+				packer.Close()
 			},
 		),
 	}
@@ -160,7 +158,7 @@ func makeBatchForTest(
 func BenchmarkTxnTableInsert(b *testing.B) {
 	ctx := context.Background()
 	mp := mpool.MustNewZero()
-	table := newTxnTableForTest(mp)
+	table := newTxnTableForTest()
 	for i, max := int64(0), int64(b.N); i < max; i++ {
 		err := table.Write(
 			ctx,
