+++ pkg/vm/engine/tae/index/zm_test.go
@@ -275,7 +275,7 @@ func TestZMNull(t *testing.T) {
 }
 
 func TestZmStringCompose(t *testing.T) {
-	packer := types.NewPacker(mpool.MustNewNoFixed("TestZmCompose"))
+	packer := types.NewPacker()
 	packer.EncodeStringType([]byte("0123456789.0123456789.0123456789."))
 	packer.EncodeInt32(42)
